Move multipliers to own module; add utils for fetching optimizers.
