+++ src/torch_constrained/__init__.py
@@ -1,2 +1,3 @@
 from .constrained_optimizer import *
 from .extra_optimizers import *
+from .multipliers import *
